requirements.txt
python-telegram-bot[job-queue]==21.6
python-dotenv==1.0.1
dateparser==1.2.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
